Move background/version/types.ts to common
