+++ src/background/version.ts
@@ -1,7 +1,7 @@
 import fs from "node:fs";
 import url from "node:url";
 import path from "node:path";
-import { Releases, VersionStatus } from "./types";
+import { Releases, VersionStatus } from "@/common/version";
 import { getAppPath, isDevelopment, isTest } from "@/background/proc/env";
 import { exists } from "@/background/helpers/file";
 import { fetch } from "@/background/helpers/http";
